refactor: move RunCommandLocationPolling class to separate file
